# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ALMA Backend API Documentation",
    "description": "Back-end API to manage users",
    "contact": {
      "name": "Rafael Ribeiro",
      "url": "https://www.wophi.be/alma/support",
      "email": "alma@wophi.be"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/user/signup": {
      "post": {
        "summary": "Create user",
        "description": "Route to register new users",
        "tags": ["User"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUp"
              },
              "examples": {
                "example1": {
                  "summary": "Example of a valid user registration",
                  "value": {
                    "firstName": "John",
                    "lastName": "Doe",
                    "socialName": "Joane Eod",
                    "username": "jd",
                    "bornDate": "1990-01-01",
                    "motherName": "Jane Doe",
                    "email": "johndoe@example.com",
                    "phone": "1234567890",
                    "password": "#Password123",
                    "passwordConfirmation": "#Password123",
                    "originChannel": "WOPHI"
                  }
                },
                "example2": {
                  "summary": "Example of required fields only",
                  "value": {
                    "firstName": "John",
                    "lastName": "Doe",
                    "motherName": "Jane Doe",
                    "phone": "1234567890",
                    "email": "johndoe@example.com",
                    "password": "#Password123",
                    "passwordConfirmation": "#Password123",
                    "originChannel": "WOPHI"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "id": "123456",
                  "personal": {
                    "id": "789012",
                    "firstName": "John",
                    "lastName": "Doe",
                    "socialName": "Joane Oed",
                    "bornDate": "1990-01-01",
                    "motherName": "Jane Doe"
                  },
                  "contact": {
                    "id": "345678",
                    "username": "jd",
                    "phone": "1234567890",
                    "email": "johndoe@example.com"
                  },
                  "security": {
                    "id": "901234",
                    "status": "PENDING_CONFIRMATION"
                  },
                  "allowedChannels": ["WOPHI"],
                  "createdAt": "2023-05-10T12:00:00Z",
                  "updatedAt": "2023-05-10T12:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": ["One or more fields are missing or invalid"],
                    "code": "bad.request",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": "invalid ip address",
                    "code": "user-service.createUser",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "example1": {
                    "summary": "field already in use by another user",
                    "value": {
                      "error": {
                        "message": "[ 'email', 'phone' ] already in use",
                        "code": "user-repository.createUser",
                        "status": true
                      },
                      "data": {}
                    }
                  },
                  "example2": {
                    "summary": "accessing new channel",
                    "value": {
                      "error": {
                        "message": "User 'user.id' registered on [ user.allowed_channels ] platform. A confirmation token was sent to 'user.contact.email' and after confirmation, use the same credentials from 'user.allowed_channels' to access [ data.originChannel ].",
                        "code": "user-service.createUser",
                        "status": true
                      },
                      "data": {}
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": "passwords do not match",
                    "code": "user-service.createUser",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },

                "examples": {
                  "example1": {
                    "summary": "repository error",
                    "value": {
                      "error": {
                        "message": "user not created",
                        "code": "user-repository.createUser",
                        "status": true
                      },
                      "data": {}
                    }
                  },
                  "example2": {
                    "summary": "service error",
                    "value": {
                      "error": {
                        "message": "failed to create user",
                        "code": "user-service.createUser",
                        "status": true
                      },
                      "data": {}
                    }
                  },
                  "example3": {
                    "summary": "email service error",
                    "value": {
                      "error": {
                        "message": "failed to send email for account confirmation",
                        "code": "email-service.sendConfirmationEmail",
                        "status": true
                      },
                      "data": {}
                    }
                  },
                  "example4": {
                    "summary": "failed to create access to new channel service",
                    "value": {
                      "error": {
                        "message": "failed to create access to new channel",
                        "code": "user-service.createUser",
                        "status": true
                      },
                      "data": {}
                    }
                  },
                  "example5": {
                    "summary": "failed to create access to new channel repository",
                    "value": {
                      "error": {
                        "message": "failed to create access to new channel",
                        "code": "user-service.createUser",
                        "status": true
                      },
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/user/filter": {
      "get": {
        "summary": "Get user by filter",
        "description": "Route to get users by filter [id | email | phone]",
        "tags": ["User"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer {token}",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "User id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "description": "User email",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "in": "query",
            "description": "User phone",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "examples": {
                  "example1": {
                    "summary": "User found",
                    "value": {
                      "id": "123456",
                      "personal": {
                        "id": "789012",
                        "firstName": "John",
                        "lastName": "Doe",
                        "socialName": "JD",
                        "bornDate": "1990-01-01",
                        "motherName": "Jane Doe",
                        "updatedAt": "2023-05-10T12:00:00Z"
                      },
                      "contact": {
                        "id": "345678",
                        "username": "johndoe",
                        "email": "johndoe@example.com",
                        "phone": "1234567890",
                        "updatedAt": "2023-05-10T12:00:00Z"
                      },
                      "security": {
                        "id": "901234",
                        "status": "ACTIVE",
                        "updatedAt": "2023-05-11T12:00:00Z"
                      },
                      "allowedChannels": ["WOPHI"],
                      "createdAt": "2023-05-10T12:00:00Z",
                      "updatedAt": "2023-05-11T12:00:00Z"
                    }
                  },
                  "example2": {
                    "summary": "User not found",
                    "value": {}
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": "Unauthorized",
                    "code": "auth-guard.jwtAuth",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": "could not get user",
                    "code": "user-repository.userByFilter",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "Get user by id",
        "description": "Route to get an user by id",
        "tags": ["User"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer {token}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "id": "123456",
                  "personal": {
                    "id": "789012",
                    "firstName": "John",
                    "lastName": "Doe",
                    "socialName": "JD",
                    "bornDate": "1990-01-01",
                    "motherName": "Jane Doe",
                    "updatedAt": "2023-05-10T12:00:00Z"
                  },
                  "contact": {
                    "id": "345678",
                    "username": "johndoe",
                    "email": "johndoe@example.com",
                    "phone": "1234567890",
                    "updatedAt": "2023-05-10T12:00:00Z"
                  },
                  "security": {
                    "id": "901234",
                    "status": "ACTIVE",
                    "updatedAt": "2023-05-11T12:00:00Z"
                  },
                  "allowedChannels": ["WOPHI"],
                  "createdAt": "2023-05-10T12:00:00Z",
                  "updatedAt": "2023-05-11T12:00:00Z"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": "Unauthorized",
                    "code": "auth-guard.jwtAuth",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": "user not found",
                    "code": "user-repository.getUserById",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          }
        }
      }
    },
    "/user/update": {
      "patch": {
        "summary": "Update user",
        "description": "Route to update user data",
        "tags": ["User"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer {token}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              },
              "examples": {
                "example1": {
                  "summary": "Example of updating password",
                  "value": {
                    "oldPassword": "password123",
                    "newPassword": "newpassword",
                    "passwordConfirmation": "newpassword"
                  }
                },
                "example2": {
                  "summary": "Example of updating username and email",
                  "value": {
                    "username": "Jd",
                    "email": "jd@mail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "id": "123456",
                  "personal": {
                    "id": "789012",
                    "firstName": "John",
                    "lastName": "Doe",
                    "socialName": "JD",
                    "bornDate": "1990-01-01",
                    "motherName": "Jane Doe",
                    "updatedAt": "2023-05-10T12:00:00Z"
                  },
                  "contact": {
                    "id": "345678",
                    "username": "jd",
                    "email": "jd@mail.com",
                    "phone": "1234567890",
                    "updatedAt": "2023-05-13T12:00:00Z"
                  },
                  "security": {
                    "id": "901234",
                    "status": "ACTIVE",
                    "updatedAt": "2023-12-11T12:00:00Z"
                  },
                  "allowedChannels": ["WOPHI"],
                  "createdAt": "2023-05-10T12:00:00Z",
                  "updatedAt": "2023-12-11T12:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "example1": {
                    "summary": "invalid user id",
                    "value": {
                      "error": {
                        "message": ["user id not found"],
                        "code": "bad.request",
                        "status": true
                      },
                      "data": {}
                    }
                  },
                  "example2": {
                    "summary": "invalid user id",
                    "value": {
                      "error": {
                        "message": [
                          "One or more fields are missing or invalid"
                        ],
                        "code": "bad.request",
                        "status": true
                      },
                      "data": {}
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": "Unauthorized",
                    "code": "auth-guard.jwtAuth",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": "invalid ip address",
                    "code": "user-service.updateUser",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": "email/phone already taken",
                    "code": "user-repository.createUser",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          },
          "422": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "example1": {
                    "summary": "missing old password propertie",
                    "value": {
                      "error": {
                        "message": "missing 'oldPassword' field",
                        "code": "user-service.updateUser",
                        "status": true
                      },
                      "data": {}
                    }
                  },
                  "example2": {
                    "summary": "validate new password",
                    "value": {
                      "error": {
                        "message": "new passwords do not match",
                        "code": "user-service.updateUser",
                        "status": true
                      },
                      "data": {}
                    }
                  },
                  "example3": {
                    "summary": "validate old password",
                    "value": {
                      "error": {
                        "message": "old passwords do not match",
                        "code": "user-repository.updateUser",
                        "status": true
                      },
                      "data": {}
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "example1": {
                    "summary": "repository internal error",
                    "value": {
                      "error": {
                        "message": "user not updated",
                        "code": "user-repository.updateUser",
                        "status": true
                      },
                      "data": {}
                    }
                  },
                  "example2": {
                    "summary": "service internal error",
                    "value": {
                      "error": {
                        "message": "failed to update user data",
                        "code": "user-service.updateUser",
                        "status": true
                      },
                      "data": {}
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/delete": {
      "delete": {
        "summary": "Delete account",
        "description": "Route to delete user account. After this operation, user will receive the 'CANCELLED' status  and after 21 days if not reactivate, the account will be permanently deleted",
        "tags": ["User"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer {token}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "id": "123456",
                  "personal": {
                    "id": "789012",
                    "firstName": "John",
                    "lastName": "Doe",
                    "socialName": "JD",
                    "bornDate": "1990-01-01",
                    "motherName": "Jane Doe",
                    "updatedAt": "2023-05-10T12:00:00Z"
                  },
                  "contact": {
                    "id": "345678",
                    "username": "jd",
                    "email": "jd@mail.com",
                    "phone": "1234567890",
                    "updatedAt": "2023-05-13T12:00:00Z"
                  },
                  "security": {
                    "id": "901234",
                    "status": "CANCELLED",
                    "updatedAt": "2023-12-11T12:00:00Z"
                  },
                  "allowedChannels": ["WOPHI"],
                  "createdAt": "2023-05-10T12:00:00Z",
                  "updatedAt": "2023-12-11T12:00:00Z"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": "Unauthorized",
                    "code": "auth-guard.jwtAuth",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": {
                    "message": "user not cancelled",
                    "code": "user-repository.deleteUser",
                    "status": true
                  },
                  "data": {}
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SignUp": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "maxLength": 125,
            "description": "First name of the user"
          },
          "lastName": {
            "type": "string",
            "maxLength": 125,
            "description": "Last name of the user"
          },
          "socialName": {
            "type": "string",
            "maxLength": 250,
            "description": "Social name of the user"
          },
          "bornDate": {
            "type": "string",
            "formate": "date-time",
            "description": "Date of birth of the user (yyyy-mm-dd)"
          },
          "motherName": {
            "type": "string",
            "maxLength": 250,
            "description": "Mother's name of the user"
          },
          "username": {
            "type": "string",
            "maxLength": 45,
            "description": "Username of the user"
          },
          "email": {
            "type": "string",
            "maxLength": 250,
            "description": "Email address of the user"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the user"
          },
          "password": {
            "type": "string",
            "minLength": "7",
            "pattern": "at least one uppercase letter, one lowercase letter and one number or symbol",
            "description": "Password of the user"
          },
          "passwordConfirmation": {
            "type": "string",
            "description": "Password confirmation"
          },
          "originChannel": {
            "type": "string",
            "enum": ["WOPHI", "LUMIN", "MIAU"],
            "description": "Origin channel of the user"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "bornDate",
          "motherName",
          "email",
          "phone",
          "password",
          "passwordConfirmation",
          "originChannel"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "personal": {
            "$ref": "#/components/schemas/Personal"
          },
          "contact": {
            "$ref": "#/components/schemas/Contact"
          },
          "security": {
            "$ref": "#/components/schemas/Security"
          },
          "allowedChannels": {
            "type": "string",
            "enum": ["WOPHI", "LUMIN", "MIAU"],
            "description": "Channels that user can access"
          },
          "createdAt": {
            "type": "string",
            "description": "Date and time of user creation"
          },
          "updatedAt": {
            "type": "string",
            "description": "Date and time of last user update"
          }
        }
      },
      "Personal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Personal information ID"
          },
          "firstName": {
            "type": "string",
            "description": "First name of the user"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the user"
          },
          "socialName": {
            "type": "string",
            "description": "Social name of the user"
          },
          "bornDate": {
            "type": "string",
            "description": "Date of birth of the user"
          },
          "motherName": {
            "type": "string",
            "description": "Mother's name of the user"
          }
        }
      },
...
      "UpdateUser": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name of the user"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the user"
          },
          "socialName": {
            "type": "string",
            "description": "Social name of the user"
          },
          "bornDate": {
            "type": "string",
            "description": "Date of birth of the user"
          },
          "motherName": {
            "type": "string",
            "description": "Mother's name of the user"
          },
          "username": {
            "type": "string",
            "description": "Username of the user"
          },
          "email": {
            "type": "string",
            "description": "Email address of the user"
          },
          "phone": {
            "type": "string",
            "description": "Phone number of the user"
          },
          "oldPassword": {
            "type": "string",
            "description": "Account old password"
          },
          "newPassword": {
            "type": "string",
            "description": "Account new password"
          },
          "passwordConfirmation": {
            "type": "string",
            "description": "Password confirmation"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING_CONFIRMATION",
              "ACTIVE",
              "BLOCKED",
              "CANCELLED",
              "BANNED"
            ],
            "description": "Status of the user"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Error message"
              },
              "code": {
                "type": "string",
                "description": "Error code"
              },
              "status": {
                "type": "boolean",
                "description": "Error status"
              }
            }
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}